Further work on entity interpolation

+++ adventure-server/src/main/java/com/adam/adventure/server/tick/ServerTickSchedulerFactory.java
@@ -1,5 +1,5 @@
 package com.adam.adventure.server.tick;
 
 public interface ServerTickSchedulerFactory {
-    ServerTickScheduler create(final int tickrate);
+    ServerTickScheduler create(final long tickrate);
 }
